[#15] MainScreen 코드 분리
